requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pillow==10.0.0
pyzbar==0.1.9
torch==2.0.1
torchvision==0.15.2
pandas==2.0.3
pyyaml==6.0
psutil==5.9.5
matplotlib==3.7.2
seaborn==0.12.2
pyserial==3.5 